+++ src/server/hx/bar.rs
@@ -1,26 +1,19 @@
 use super::gallery::render;
-use super::ServerError;
+use super::render::Response;
 use super::State;
+use actix_web::get;
 use actix_web::web::{Data, Path, Query};
-use actix_web::{get, HttpResponse};
 use sqlx::{Pool, Sqlite};
 
 #[get("/favorite")]
-async fn favorite(
-    pool: Data<Pool<Sqlite>>,
-    state: Query<State>,
-) -> Result<HttpResponse, ServerError> {
+async fn favorite(pool: Data<Pool<Sqlite>>, state: Query<State>) -> Response {
     let mut state = state.into_inner();
     state.favorite = state.favorite.or(Some(false)).map(|b| !b);
     render(&pool, state).await
 }
 
 #[get("/year/{year}")]
-async fn year(
-    pool: Data<Pool<Sqlite>>,
-    state: Query<State>,
-    path: Path<usize>,
-) -> Result<HttpResponse, ServerError> {
+async fn year(pool: Data<Pool<Sqlite>>, state: Query<State>, path: Path<usize>) -> Response {
     let mut state = state.into_inner();
     let year = path.into_inner();
     if state.year == Some(year) {
@@ -33,11 +26,7 @@ async fn year(
 }
 
 #[get("/month/{month}")]
-async fn month(
-    pool: Data<Pool<Sqlite>>,
-    state: Query<State>,
-    path: Path<u8>,
-) -> Result<HttpResponse, ServerError> {
+async fn month(pool: Data<Pool<Sqlite>>, state: Query<State>, path: Path<u8>) -> Response {
     let mut state = state.into_inner();
     let month = path.into_inner();
     assert!(state.year.is_some());
@@ -50,6 +39,6 @@ async fn month(
 }
 
 #[get("/bar/clear")]
-async fn clear(pool: Data<Pool<Sqlite>>) -> Result<HttpResponse, ServerError> {
+async fn clear(pool: Data<Pool<Sqlite>>) -> Response {
     render(&pool, State::default()).await
 }
