+++ src/server/hx/preview.rs
@@ -1,11 +1,11 @@
+use std::str::FromStr;
+
 use crate::db;
 
+use super::render::{Response, ServerError, TemplatedResponse};
 use super::Media;
-use super::ServerError;
-use super::TemplatedResponse;
 use actix_web::web;
 use actix_web::web::Data;
-use actix_web::HttpResponse;
 use actix_web::{delete, get, put};
 use askama::Template;
 use sqlx::{Pool, Sqlite};
@@ -18,12 +18,9 @@ struct HxRoot<'a> {
 }
 
 #[get("/{idx}")]
-async fn root(
-    pool: Data<Pool<Sqlite>>,
-    uuid: web::Path<Uuid>,
-) -> Result<HttpResponse, ServerError> {
+async fn root(pool: Data<Pool<Sqlite>>, uuid: web::Path<Uuid>) -> Response {
     let media = Media {
-        uuid: "123123".into(),
+        uuid: Uuid::from_str("9be1c561-5245-42a3-af22-b0c77136665f").unwrap(),
         url: "http://stufur:1337/assets/media/tota_myndir/2018/20180723_183916.jpg".into(),
         thumbnail: "http://stufur:1337/assets/thumbnails/1601707f-b75e-3640-91e4-0c4331ec7f6e.webp"
             .into(),
@@ -39,16 +36,13 @@ async fn root(
 }
 
 #[put("/{idx}/favorite")]
-async fn favorite(
-    pool: Data<Pool<Sqlite>>,
-    uuid: web::Path<Uuid>,
-) -> Result<HttpResponse, ServerError> {
+async fn favorite(pool: Data<Pool<Sqlite>>, uuid: web::Path<Uuid>) -> Response {
     db::media_favorite(&pool, &uuid, true)
         .await
         .map_err(ServerError::DBError)?;
 
     let media = Media {
-        uuid: "123123".into(),
+        uuid: Uuid::from_str("9be1c561-5245-42a3-af22-b0c77136665f").unwrap(),
         url: "http://stufur:1337/assets/media/tota_myndir/2018/20180723_183916.jpg".into(),
         thumbnail: "http://stufur:1337/assets/thumbnails/1601707f-b75e-3640-91e4-0c4331ec7f6e.webp"
             .into(),
@@ -64,10 +58,7 @@ async fn favorite(
 }
 
 #[delete("/{idx}")]
-async fn archive(
-    pool: Data<Pool<Sqlite>>,
-    uuid: web::Path<Uuid>,
-) -> Result<HttpResponse, ServerError> {
+async fn archive(pool: Data<Pool<Sqlite>>, uuid: web::Path<Uuid>) -> Response {
     db::media_archive(&pool, &uuid, true)
         .await
         .map_err(ServerError::DBError)?;
